compute: remove the correction v1 buffer
The `enable_compute_correction_v2` flag has served `true` in production for
every build at or beyond 26.7.0-rc.1, so the v1 correction buffer is dead code
in every deployment we support. This change deletes it along with the flag that
selected between the two implementations.

`CorrectionV2` becomes the only `Correction`, and its file moves to
`sink/correction.rs`. The logging and metrics helpers that both implementations
shared move to `sink/correction/logging.rs`, which keeps the buffer file from
growing further. `Correction::new` now reads the chain proportionality and chunk
size from the `ConfigSet`, matching the constructor the sink previously called
on the enum wrapper; `Correction::with_params` takes the two values directly for
tests and benchmarks.

Removing v1 also removes its only consumer of `ConsolidatingVec`, so that type
and the `consolidating_vec_growth_dampener` dyncfg that tuned it are gone as
well. Both flag keys move to the LaunchDarkly consistency check's stale list,
because the last published release still synchronizes them.

The `equivalence_with_v1` unit test compared the two implementations step by
step. It is replaced by `equivalence_with_reference`, which runs the same
upsert-and-feedback workload against a naive in-test buffer that keeps every
update in a flat vector, plus an assertion that each step emits something so the
comparison cannot pass vacuously. The correction benchmark loses its
version-dispatch wrapper and now measures the single implementation.

The dyncfg keys `compute_correction_v2_chain_proportionality` and
`compute_correction_v2_chunk_size` keep their names. Renaming them would drop
the production overrides currently set against those keys.
