+++ misc/python/materialize/mzcompose/__init__.py
@@ -96,7 +96,6 @@ def get_minimal_system_parameters(
         "enable_cast_elimination": "true",
         "enable_coalesce_case_transform": "true",
         "enable_columnation_lgalloc": "false",
-        "enable_compute_correction_v2": "true",
         "enable_compute_logical_backpressure": "true",
         "enable_connection_validation_syntax": "true",
         "enable_create_table_from_source": "true",
@@ -643,7 +642,6 @@ UNINTERESTING_SYSTEM_PARAMETERS = [
     "compute_dataflow_max_inflight_bytes_cc",
     "compute_flat_map_fuel",
     "compute_temporal_bucketing_summary",
-    "consolidating_vec_growth_dampener",
     "copy_to_s3_parquet_row_group_file_ratio",
     "copy_to_s3_arrow_builder_buffer_ratio",
     "copy_to_s3_multipart_part_size_bytes",
