+++ misc/python/materialize/parallel_workload/action.py
@@ -3187,7 +3187,6 @@ class FlipFlagsAction(Action):
             "persist_blob_hedged_get_warm_interval",
             "enable_compute_half_join2",
             "enable_mz_join_core",
-            "enable_compute_correction_v2",
             "linear_join_yielding",
             "enable_lgalloc",
             "enable_lgalloc_eager_reclamation",
@@ -3207,7 +3206,6 @@ class FlipFlagsAction(Action):
             "compute_dataflow_max_inflight_bytes_cc",
             "subscribe_max_buffered_bytes",
             "compute_flat_map_fuel",
-            "consolidating_vec_growth_dampener",
             "compute_hydration_concurrency",
             "copy_to_s3_parquet_row_group_file_ratio",
             "copy_to_s3_arrow_builder_buffer_ratio",
