+++ src/compute/src/sink.rs
@@ -12,10 +12,6 @@ mod copy_to_s3_oneshot;
 pub mod correction;
 #[cfg(not(feature = "bench"))]
 mod correction;
-#[cfg(feature = "bench")]
-pub mod correction_v2;
-#[cfg(not(feature = "bench"))]
-mod correction_v2;
 mod materialized_view;
 mod materialized_view_v2;
 mod metric_sink;
