+++ src/compute/src/sink/materialized_view.rs
@@ -155,7 +155,8 @@ use crate::compute_state::ComputeState;
 use crate::render::StartSignal;
 use crate::render::errors::DataflowErrorSer;
 use crate::render::sinks::SinkRender;
-use crate::sink::correction::{ChannelLogging, Correction, CorrectionLogger};
+use crate::sink::correction::Correction;
+use crate::sink::correction::logging::{ChannelLogging, CorrectionLogger};
 use crate::sink::materialized_view_v2;
 use crate::sink::refresh::apply_refresh;
