+++ test/launchdarkly-flag-consistency/mzcompose.py
@@ -459,9 +459,11 @@ KNOWN_STALE_LD_FLAGS: set[str] = set("""
     clusterd_malloc_conf
     column_paged_batcher_spill_worker_count
     column_paged_batcher_use_pool
+    consolidating_vec_growth_dampener
     constraint_based_timestamp_selection
     enable_0dt_deployment
     enable_aws_msk_iam_auth
+    enable_compute_correction_v2
     enable_consolidate_after_union_negate
     enable_continual_task_builtins
     enable_continual_task_transform
@@ -532,7 +534,6 @@ INTENTIONAL_LD_OVERRIDES: set[str] = {
     # flag to a permanent category above.
     "compute_subscribe_snapshot_optimization",
     "enable_cast_elimination",
-    "enable_compute_correction_v2",
     "enable_compute_error_distinct",
     "enable_compute_temporal_bucketing",
     "enable_new_outer_join_lowering",
@@ -557,7 +558,6 @@ KNOWN_CROSS_ENV_DIVERGENCES: set[str] = set("""
     enable_cluster_schedule_refresh
     enable_column_paged_batcher
     enable_column_paged_batcher_spill
-    enable_compute_correction_v2
     enable_eager_delta_joins
     enable_index_options
     enable_join_prioritize_arranged
